With --store local
Based on a suggestion from Robert Hensing [1], add --store local to the
nix-daemon invocation. This revealed further issues that I chased for a
little while. I'm now running into issues with the store that's mounted
into the VM being read-only, which wouldn't happen on a full machine,
but I'll worry about that later.

[1]: https://functional.cafe/@roberth/114319211674977879